modified FacilityDepartments to FacilityDepartment, with single facility/department per table entry
